Rename React JSX modules to use .jsx extension
