+++ frontend/src/components/CreatorCard.jsx
