+++ frontend/src/pages/CollectionsPage.jsx
